#22 moved Authentication related tests to own test-suite
